Split Resources syntax generation code into partial file
